Renamet navn på testfiler.
